backend/requirements.txt
opencv-python
mediapipe
numpy
pillow
flask
flask-cors
torch
torchvision
face-recognition
pandas
openpyxl
